Update LaravelBatch.php
Modified the existing update function as a requirement of matching with two in columns values. This may be helpful as a quick requirement. I didn't create up the dynamic numbers for matching but only one extra index.

WHEN (CONDITION1 AND CONDITION2) THEN VALUE.

USING THE BATCH HELPER IT WILL BE LIKE: -
$updateIndex1 = 'id';
$updateIndex1 = 'user_id';
batch()->updateWithTwoIndex($userTickerMappingModel, $updatedTicketData,$updateIndex1, $updateIndex2);